Fix cyclic dependencies and implement sync_submission_review
